requirements.txt
absl-py==1.4.0
blis==0.4.1
cached-property==1.5.2
cachetools==4.2.4
catalogue==2.0.10
certifi==2021.5.30
charset-normalizer==2.0.12
click==8.0.4
confection==0.1.4
contextvars==2.4
cycler==0.11.0
cymem==2.0.8
dataclasses 
en-core-web-md==2.1.0
en-core-web-sm==2.1.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
grpcio==1.48.2
h5py==3.1.0
idna==3.7
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
Jinja2==3.0.3
kiwisolver==1.3.1
langcodes==3.3.0
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
murmurhash==1.0.10
numpy==1.19.5
oauthlib==3.2.2
olefile==0.46
packaging==21.3
pathy==0.10.3
Pillow==8.4.0
plac==1.1.3
preshed==3.0.9
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pydantic==1.9.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
requests==2.27.1
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
smart-open==6.4.0
spacy==2.2.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==1.0.7
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thinc==7.3.1
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tqdm==4.64.1
typer==0.9.4
typing_extensions==4.1.1
urllib3==1.26.18
wasabi==0.10.1
Werkzeug==2.0.3
zipp==3.6.0
